// src/api/generated/models/userRole.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRole = {
  caregiver: 'caregiver',
  district_focal_person: 'district_focal_person',
  ncda_admin: 'ncda_admin',
  ecd_director: 'ecd_director',
} as const;